Added playlist support
Now you can add the whole playlist in the queue. If the playlist is longer than 100 elements, it will not include the ones after 100. Look through ytpl documentation to accept every element.
The whole search portion got made into a function so that it will be used by other commands, playskip & playtop.