Added playskip and playtop functionality
You can now add a song at the top of the queue and immediately play it with playskip
Playtop is the same but it won't skip the current playing song.

Fixes on now_playing.
Since we now have playtop and skip, their position index won't be in the end of the queue as it was per default. I added a position parameter. This parameter also helped on showing the correct thumbnail image on playlists, so that it will show the first song's and not the last.